Moved useCheckUser to the dedicated customHooks folder and changed all references accordingly.
