+++ frontend/src/DeleteBookstoreModal.jsx
@@ -1,4 +1,4 @@
-import useCheckUser from "./useCheckUser";
+import useCheckUser from "./customHooks/useCheckUser";
 
 function DeleteBookstoreModal({ row, closeDeleteModal, pageIndex, globalFilter }) {
   useCheckUser();
