+++ frontend/src/DeleteCategoryModal.jsx
@@ -1,4 +1,4 @@
-import useCheckUser from "./useCheckUser";
+import useCheckUser from "./customHooks/useCheckUser";
 
 function DeleteCategoryModal( {row, closeDeleteModal, pageIndex, globalFilter }) {
   useCheckUser();
