+++ frontend/src/customHooks/useCheckUser.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import UserContext from './UserContext';
+import UserContext from '../UserContext';
 
 function useCheckUser() {
   const { user, fetchUser } = useContext(UserContext);
